+++ src/fixate/drivers/__init__.py
@@ -1,8 +1,4 @@
-try:
-    from typing import Protocol
-except ImportError:
-    # Protocol added in python 3.8
-    from typing_extensions import Protocol
+from typing import Protocol
 
 import pubsub.pub