+++ docs/source/release-notes.rst
@@ -2,6 +2,20 @@
 Release Notes
 ==================================
 *************
+Version 0.6.2
+*************
+Release Date 05/06/24
+
+New Features
+############
+
+Improvements
+############
+- Add logging of computer name to log file.
+- Update DMM driver with optional measurement delay to improve DMM model compatibility.
+
+
+*************
 Version 0.6.1
 *************
 Release Date 22/05/2023